feat: upgrade to Angular 13
Add eslint
Add core-js
Enable Ivy
Enable Aot
Fix polyfills
Fix lint errors
Fix template compilation errors
Fix example-app build
Fix docs-app build
Fix cyclical component references (https://github.com/angular/angular/issues/43227#issuecomment-904173738)

BREAKING CHANGE: Library is now build with 'partial' Ivy
