+++ .gitignore
@@ -68,3 +68,5 @@ testResults
 e2eResults
 
 e2e/dist
+
+.angular
