+++ package.json
@@ -38,7 +38,7 @@
     "copy:files:win": "copy README.md .\\dist\\angular-tree-component && copy LICENSE .\\dist\\angular-tree-component && xcopy .\\projects\\angular-tree-component\\src\\lib\\angular-tree-component.css .\\dist\\angular-tree-component\\css\\",
     "doc": "typedoc --module commonjs --out doc lib/defs/api.ts",
     "clean:all": "npm run clean:dist && rimraf node_modules && npm cache clean",
-    "lint": "tslint lib/**/*.ts",
+    "lint": "npx eslint",
     "prepublish:lib:win": "npm run build:lib && npm run copy:files:win",
     "start:example-app": "ng serve",
     "start:example-app:es5": "ng serve --configuration=es5",
@@ -61,56 +61,60 @@
     "Adam Klein <adam@500tech.com>"
   ],
   "dependencies": {
-    "@angular/animations": "^11.2.0",
-    "@angular/cdk": "^11.2.0",
-    "@angular/forms": "^11.2.0",
-    "@angular/material": "^11.2.0",
+    "@angular/animations": "^13.3.8",
+    "@angular/cdk": "^13.3.7",
+    "@angular/forms": "^13.3.8",
+    "@angular/material": "^13.3.7",
+    "core-js": "^3.22.5",
     "mobx": "~4.14.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.0",
-    "@angular-devkit/build-ng-packagr": "~0.1002.0",
-    "@angular-devkit/core": "^11.2.0",
-    "@angular/cli": "^11.2.0",
-    "@angular/common": "^11.2.0",
-    "@angular/compiler": "^11.2.0",
-    "@angular/compiler-cli": "^11.2.0",
-    "@angular/core": "^11.2.0",
-    "@angular/elements": "^11.2.0",
-    "@angular/platform-browser": "^11.2.0",
-    "@angular/platform-browser-dynamic": "^11.2.0",
-    "@angular/platform-server": "^11.2.0",
-    "@angular/router": "^11.2.0",
+    "@angular-devkit/architect": "^0.1303.5",
+    "@angular-devkit/build-angular": "^13.3.5",
+    "@angular-devkit/core": "^13.3.5",
+    "@angular-eslint/builder": "^13.2.1",
+    "@angular-eslint/schematics": "^13.2.1",
+    "@angular-eslint/template-parser": "^13.2.1",
+    "@angular/cli": "^13.3.5",
+    "@angular/common": "^13.3.8",
+    "@angular/compiler": "^13.3.8",
+    "@angular/compiler-cli": "^13.3.8",
+    "@angular/core": "^13.3.8",
+    "@angular/elements": "^13.3.8",
+    "@angular/platform-browser": "^13.3.8",
+    "@angular/platform-browser-dynamic": "^13.3.8",
+    "@angular/platform-server": "^13.3.8",
+    "@angular/router": "^13.3.8",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.7",
     "@types/rx": "2.5.34",
     "@types/webpack": "^1.12.29",
-    "codelyzer": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "eslint": "^8.15.0",
     "html-dnd": "1.2.0",
     "http-server": "0.9.0",
     "husky": "^4.2.3",
-    "lint-staged": "^10.0.7",
-    "ng-packagr": "^11.2.1",
-    "nightwatch": "0.9.12",
-    "prettier": "^1.19.1",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.1.1",
+    "karma": "~6.3.20",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
+    "lint-staged": "^10.0.7",
+    "ng-packagr": "^13.3.1",
+    "nightwatch": "0.9.12",
+    "prettier": "^1.19.1",
     "rimraf": "^2.5.1",
     "rxjs": "^6.5.3",
-    "testcafe": "^1.8.8",
-    "tsickle": "^0.38.1",
-    "tslib": "^2.0.0",
-    "tslint": "~6.1.0",
+    "testcafe": "^1.18.6",
+    "tslib": "^2.4.0",
     "typedoc": "^0.3.12",
-    "typescript": "~4.1.5",
+    "typescript": "~4.6.4",
     "uuid": "^8.2.0",
-    "zone.js": "^0.10.2"
+    "zone.js": "^0.11.5"
   },
   "repository": "https://github.com/CirclonGroup/angular-tree-component.git",
   "bugs": {
