+++ projects/angular-tree-component/ng-package.json
@@ -3,11 +3,7 @@
   "dest": "../../dist/angular-tree-component",
   "assets": ["src/lib/angular-tree-component.css"],
   "lib": {
-    "entryFile": "src/public-api.ts",
-    "umdModuleIds": {
-      "mobx": "mobx",
-      "lodash-es": "lodash-es"
-    }
+    "entryFile": "src/public-api.ts"
   },
-  "whitelistedNonPeerDependencies": ["core-js", "lodash-es", "mobx"]
+  "allowedNonPeerDependencies": ["core-js", "lodash-es", "mobx"]
 }
