+++ projects/angular-tree-component/src/lib/angular-tree-component.module.ts
@@ -20,12 +20,10 @@ import { TreeDraggedElement } from './models/tree-dragged-element.model';
 import { TreeVirtualScroll } from './models/tree-virtual-scroll.model';
 import { LoadingComponent } from './components/loading.component';
 import { TreeComponent } from './components/tree.component';
-import { TreeNodeComponent } from './components/tree-node.component';
-import { TreeNodeContent } from './components/tree-node-content.component';
-import { TreeNodeDropSlot } from './components/tree-node-drop-slot.component';
+import { TreeNodeContentComponent } from './components/tree-node-content.component';
+import { TreeNodeDropSlotComponent } from './components/tree-node-drop-slot.component';
 import { TreeNodeExpanderComponent } from './components/tree-node-expander.component';
-import { TreeNodeChildrenComponent } from './components/tree-node-children.component';
-import { TreeNodeCollectionComponent } from './components/tree-node-collection.component';
+import { TreeNodeChildrenComponent, TreeNodeCollectionComponent, TreeNodeComponent } from './components/tree-node-collection.component';
 import { TreeNodeWrapperComponent } from './components/tree-node-wrapper.component';
 import { TreeViewportComponent } from './components/tree-viewport.component';
 import { TreeNodeCheckboxComponent } from './components/tree-node-checkbox.component';
@@ -37,13 +35,13 @@ import { TreeAnimateOpenDirective } from './directives/tree-animate-open.directi
   declarations: [
     TreeComponent,
     TreeNodeComponent,
-    TreeNodeContent,
+    TreeNodeContentComponent,
     LoadingComponent,
     TreeDropDirective,
     TreeDragDirective,
     TreeNodeExpanderComponent,
     TreeNodeChildrenComponent,
-    TreeNodeDropSlot,
+    TreeNodeDropSlotComponent,
     TreeNodeCollectionComponent,
     TreeViewportComponent,
     TreeNodeWrapperComponent,
@@ -54,13 +52,13 @@ import { TreeAnimateOpenDirective } from './directives/tree-animate-open.directi
   exports: [
     TreeComponent,
     TreeNodeComponent,
-    TreeNodeContent,
+    TreeNodeContentComponent,
     LoadingComponent,
     TreeDropDirective,
     TreeDragDirective,
     TreeNodeExpanderComponent,
     TreeNodeChildrenComponent,
-    TreeNodeDropSlot,
+    TreeNodeDropSlotComponent,
     TreeNodeCollectionComponent,
     TreeViewportComponent,
     TreeNodeWrapperComponent,
@@ -90,14 +88,15 @@ export {
   TreeComponent,
   TreeNodeComponent,
   TreeNodeWrapperComponent,
-  TreeNodeContent,
+  TreeNodeContentComponent as TreeNodeContent,
   TreeDropDirective,
   TreeDragDirective,
   TreeNodeExpanderComponent,
   TreeNodeChildrenComponent,
-  TreeNodeDropSlot,
+  TreeNodeDropSlotComponent as TreeNodeDropSlot,
   TreeNodeCollectionComponent,
   TreeViewportComponent,
   TreeNodeCheckboxComponent,
-  ITreeState
+  ITreeState,
+  TreeMobxAutorunDirective
 };
