+++ projects/angular-tree-component/src/lib/components/tree-node-collection.component.ts
@@ -1,15 +1,8 @@
-import {
-  Component,
-  Input,
-  ViewEncapsulation,
-  OnInit,
-  OnDestroy
-} from '@angular/core';
-import { reaction } from 'mobx';
-import { observable, computed, action } from '../mobx-angular/mobx-proxy';
-import { TreeVirtualScroll } from '../models/tree-virtual-scroll.model';
-import { TreeNode } from '../models/tree-node.model';
-import { TreeModel } from '../models/tree.model';
+import { Component, ViewEncapsulation, Input, OnInit, OnDestroy } from "@angular/core";
+import { computed, action, reaction, observable } from "mobx";
+import { TreeNode } from "../models/tree-node.model";
+import { TreeVirtualScroll } from "../models/tree-virtual-scroll.model";
+import { TreeModel } from "../models/tree.model";
 
 @Component({
   selector: 'tree-node-collection',
@@ -95,4 +88,99 @@ export class TreeNodeCollectionComponent implements OnInit, OnDestroy {
   trackNode(index, node) {
     return node.id;
   }
+}
+
+@Component({
+  selector: 'tree-node-children',
+  encapsulation: ViewEncapsulation.None,
+  styles: [],
+  template: `
+    <ng-container *treeMobxAutorun="{ dontDetach: true }">
+      <div
+        [class.tree-children]="true"
+        [class.tree-children-no-padding]="node.options.levelPadding"
+        *treeAnimateOpen="
+          node.isExpanded;
+          speed: node.options.animateSpeed;
+          acceleration: node.options.animateAcceleration;
+          enabled: node.options.animateExpand
+        "
+      >
+        <tree-node-collection
+          *ngIf="node.children"
+          [nodes]="node.children"
+          [templates]="templates"
+          [treeModel]="node.treeModel"
+        >
+        </tree-node-collection>
+        <tree-loading-component
+          [style.padding-left]="node.getNodePadding()"
+          class="tree-node-loading"
+          *ngIf="!node.children"
+          [template]="templates.loadingTemplate"
+          [node]="node"
+        ></tree-loading-component>
+      </div>
+    </ng-container>
+  `
+})
+export class TreeNodeChildrenComponent {
+  @Input() node: TreeNode;
+  @Input() templates: any;
+}
+
+@Component({
+  selector: 'TreeNode, tree-node',
+  encapsulation: ViewEncapsulation.None,
+  styles: [],
+  template: `
+    <ng-container *treeMobxAutorun="{ dontDetach: true }">
+      <div
+        *ngIf="!templates.treeNodeFullTemplate"
+        [class]="node.getClass()"
+        [class.tree-node]="true"
+        [class.tree-node-expanded]="node.isExpanded && node.hasChildren"
+        [class.tree-node-collapsed]="node.isCollapsed && node.hasChildren"
+        [class.tree-node-leaf]="node.isLeaf"
+        [class.tree-node-active]="node.isActive"
+        [class.tree-node-focused]="node.isFocused"
+      >
+        <tree-node-drop-slot
+          *ngIf="index === 0"
+          [dropIndex]="node.index"
+          [node]="node.parent"
+        ></tree-node-drop-slot>
+
+        <tree-node-wrapper
+          [node]="node"
+          [index]="index"
+          [templates]="templates"
+        ></tree-node-wrapper>
+
+        <tree-node-children
+          [node]="node"
+          [templates]="templates"
+        ></tree-node-children>
+        <tree-node-drop-slot
+          [dropIndex]="node.index + 1"
+          [node]="node.parent"
+        ></tree-node-drop-slot>
+      </div>
+      <ng-container
+        [ngTemplateOutlet]="templates.treeNodeFullTemplate"
+        [ngTemplateOutletContext]="{
+          $implicit: node,
+          node: node,
+          index: index,
+          templates: templates
+        }"
+      >
+      </ng-container>
+    </ng-container>
+  `
+})
+export class TreeNodeComponent {
+  @Input() node: TreeNode;
+  @Input() index: number;
+  @Input() templates: any;
 }
