--- projects/angular-tree-component/src/lib/components/tree-node.component.ts
@@ -1,62 +1,0 @@
-import {
-  Component,
-  Input,
-  ViewEncapsulation,
-} from '@angular/core';
-import { TreeNode } from '../models/tree-node.model';
-
-@Component({
-  selector: 'TreeNode, tree-node',
-  encapsulation: ViewEncapsulation.None,
-  styles: [],
-  template: `
-    <ng-container *treeMobxAutorun="{ dontDetach: true }">
-      <div
-        *ngIf="!templates.treeNodeFullTemplate"
-        [class]="node.getClass()"
-        [class.tree-node]="true"
-        [class.tree-node-expanded]="node.isExpanded && node.hasChildren"
-        [class.tree-node-collapsed]="node.isCollapsed && node.hasChildren"
-        [class.tree-node-leaf]="node.isLeaf"
-        [class.tree-node-active]="node.isActive"
-        [class.tree-node-focused]="node.isFocused"
-      >
-        <tree-node-drop-slot
-          *ngIf="index === 0"
-          [dropIndex]="node.index"
-          [node]="node.parent"
-        ></tree-node-drop-slot>
-
-        <tree-node-wrapper
-          [node]="node"
-          [index]="index"
-          [templates]="templates"
-        ></tree-node-wrapper>
-
-        <tree-node-children
-          [node]="node"
-          [templates]="templates"
-        ></tree-node-children>
-        <tree-node-drop-slot
-          [dropIndex]="node.index + 1"
-          [node]="node.parent"
-        ></tree-node-drop-slot>
-      </div>
-      <ng-container
-        [ngTemplateOutlet]="templates.treeNodeFullTemplate"
-        [ngTemplateOutletContext]="{
-          $implicit: node,
-          node: node,
-          index: index,
-          templates: templates
-        }"
-      >
-      </ng-container>
-    </ng-container>
-  `
-})
-export class TreeNodeComponent {
-  @Input() node: TreeNode;
-  @Input() index: number;
-  @Input() templates: any;
-}
