+++ projects/docs-app/src/app/examples/columns-example/columns/columns.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { ITreeOptions } from 'angular-tree-component';
 
 @Component({
@@ -6,7 +6,7 @@ import { ITreeOptions } from 'angular-tree-component';
   templateUrl: './columns.component.html',
   styleUrls: ['./columns.component.scss']
 })
-export class ColumnsComponent implements OnInit {
+export class ColumnsComponent {
 
   nodes = [
     {
@@ -42,10 +42,4 @@ export class ColumnsComponent implements OnInit {
   };
 
   columns = ['city', 'zipCode'];
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
 }
