+++ projects/docs-app/src/app/examples/crud-example/crud-example.component.ts
@@ -1,15 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-crud-example',
   templateUrl: './crud-example.component.html',
   styleUrls: ['./crud-example.component.scss']
 })
-export class CrudExampleComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
+export class CrudExampleComponent { }
