+++ projects/docs-app/src/app/examples/load-more-example/load-more-example.component.ts
@@ -1,15 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-load-more-example',
   templateUrl: './load-more-example.component.html',
   styleUrls: ['./load-more-example.component.scss']
 })
-export class LoadMoreExampleComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
+export class LoadMoreExampleComponent { }
