+++ projects/docs-app/src/styles.scss
@@ -5,7 +5,7 @@
 @import "styles/scrollbar";
 @import "styles/alert";
 @import "styles/images";
-@import "~@angular/material/prebuilt-themes/indigo-pink.css";
+@import "@angular/material/prebuilt-themes/indigo-pink.css";
 @import "styles/angular-tree-component.css";
 
 // primary: #355677
